+++ src/components/__tests__/advanced-marker.test.tsx
@@ -148,8 +148,13 @@ describe('map and marker-library loaded', () => {
       .get(google.maps.marker.AdvancedMarkerElement)
       .at(0) as google.maps.marker.AdvancedMarkerElement;
 
-    expect(marker.content?.firstChild).toHaveClass('classname-test');
-    expect(marker.content?.firstChild).toHaveStyle('width: 200px');
+    const advancedMarkerWithClass = (
+      marker.content as HTMLElement
+    ).querySelector('.classname-test');
+
+    expect(advancedMarkerWithClass).toBeTruthy();
+    expect(advancedMarkerWithClass).toHaveStyle('width: 200px');
+
     expect(
       queryByTestId(marker.content as HTMLElement, 'marker-content')
     ).toBeTruthy();