+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [1.4.0](https://github.com/visgl/react-google-maps/compare/v1.3.0...v1.4.0) (2024-10-24)
+
+
+### Features
+
+* add `channel` prop to APIProvider ([#584](https://github.com/visgl/react-google-maps/issues/584)) ([6aa38e5](https://github.com/visgl/react-google-maps/commit/6aa38e52a2cf0cc856167489b879871622c74ea8))
+
+
+### Bug Fixes
+
+* adjust advanced marker markup to fix anchoring & collision behavior ([#577](https://github.com/visgl/react-google-maps/issues/577)) ([97a98b2](https://github.com/visgl/react-google-maps/commit/97a98b2a04b896a892351a178fecafa665c03113))
+
+## [1.3.0](https://github.com/visgl/react-google-maps/compare/v1.2.0...v1.3.0) (2024-09-30)
+
+
+### Features
+
+* add new `onError` prop for api provider ([#541](https://github.com/visgl/react-google-maps/issues/541)) ([bbe5709](https://github.com/visgl/react-google-maps/commit/bbe5709e1420e7260234b11a7749ed9b7804e9b7))
+
+
+### Bug Fixes
+
+* avoid unnecessary state-updates in api-provider ([#551](https://github.com/visgl/react-google-maps/issues/551)) ([46068c9](https://github.com/visgl/react-google-maps/commit/46068c9b3f3e930d464e2314181e2f6ed32a9aa7))
+
+## [1.2.0](https://github.com/visgl/react-google-maps/compare/v1.1.3...v1.2.0) (2024-09-16)
+
+
+### Features
+
+* add hover events and anchor points to advanced markers ([#472](https://github.com/visgl/react-google-maps/issues/472)) ([cc4a397](https://github.com/visgl/react-google-maps/commit/cc4a397f0ed2af12a28c21db6afad3a946527131))
+
 ## [1.1.3](https://github.com/visgl/react-google-maps/compare/v1.1.2...v1.1.3) (2024-09-08)